# api/docs/swagger.yaml
definitions:
  models.ApiError:
    properties:
      code:
        type: string
      msg:
        type: string
    type: object
...
  models.DataPoint:
    properties:
      data:
        description: The actual data payload (type depends on DataType)
      dataType:
        description: 'One of: circuits, generatorStats, prodStats, factoryStats, sinkStats'
        type: string
      gameTimeId:
        description: Game time in seconds when data was captured
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/models.ApiError'
        type: array
    type: object
...
  models.HistoryChunk:
    properties:
      dataType:
        description: The data type requested
        type: string
      latestId:
        description: Highest GameTimeID in the chunk (for client tracking)
        type: integer
      points:
        description: Array of data points, ordered by GameTimeID ascending
        items:
          $ref: '#/definitions/models.DataPoint'
        type: array
      saveName:
        description: The save name these points belong to
        type: string
    type: object
  models.HistorySavesResponse:
    properties:
      currentSave:
        description: Currently active save name for the session
        type: string
      saveNames:
        description: List of save names that have history data
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
paths:
...
  /v1/sessions/{id}/history:
    get:
      consumes:
      - application/json
      description: Returns all save names that have historical data for this session
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of save names with history
          schema:
            $ref: '#/definitions/models.HistorySavesResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List save names with history
      tags:
      - History
  /v1/sessions/{id}/history/{dataType}:
    get:
      consumes:
      - application/json
      description: Retrieves historical data points for the specified data type. Data
        is returned in ascending order by game-time ID. Use the `since` parameter
        for incremental fetching.
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Data type to retrieve history for
        enum:
        - circuits
        - generatorStats
        - prodStats
        - factoryStats
        - sinkStats
        in: path
        name: dataType
        required: true
        type: string
      - description: Save name to query (defaults to current save)
        in: query
        name: saveName
        type: string
      - description: Only return data points with gameTimeId greater than this value
          (default 0)
        in: query
        name: since
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Historical data chunk
          schema:
            $ref: '#/definitions/models.HistoryChunk'
        "400":
          description: Invalid parameters
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get historical data for a session
      tags:
      - History
...
swagger: "2.0"